Move common to gdb
