+++ numba_dpex/tests/debugging/test_breakpoints.py
@@ -13,7 +13,7 @@ import pytest
 
 from numba_dpex.tests._helper import skip_no_gdb
 
-from .common import breakpoint_by_function, breakpoint_by_mark, setup_breakpoint
+from .gdb import breakpoint_by_function, breakpoint_by_mark, setup_breakpoint
 
 pytestmark = skip_no_gdb
