+++ numba_dpex/tests/debugging/test_common.py
@@ -10,7 +10,7 @@ import pytest
 
 from numba_dpex.tests._helper import skip_no_gdb
 
-from .common import breakpoint_by_function, breakpoint_by_mark, setup_breakpoint
+from .gdb import breakpoint_by_function, breakpoint_by_mark, setup_breakpoint
 
 pytestmark = skip_no_gdb
