+++ numba_dpex/tests/debugging/test_info.py
@@ -13,7 +13,7 @@ import pytest
 
 from numba_dpex.tests._helper import skip_no_gdb
 
-from .common import setup_breakpoint
+from .gdb import setup_breakpoint
 from .test_breakpoints import side_by_side_breakpoint
 
 pytestmark = skip_no_gdb
